Merge Dylan > Master

+++ Project1/Grid.cpp
@@ -6,7 +6,7 @@ Grid::Grid() {
     cells.resize(GRID_HEIGHT, std::vector<Cell>(GRID_WIDTH, { true, {0, 0}, sf::RectangleShape(sf::Vector2f(CELL_SIZE, CELL_SIZE)) }));
     for (int y = 0; y < GRID_HEIGHT; ++y) {
         for (int x = 0; x < GRID_WIDTH; ++x) {
-            cells[y][x].position = sf::Vector2f(x * CELL_SIZE, y * CELL_SIZE);
+            cells[y][x].position = sf::Vector2f((float)(x * CELL_SIZE), (float)(y * CELL_SIZE));
             cells[y][x].shape.setPosition(cells[y][x].position);
             cells[y][x].shape.setFillColor(sf::Color::Transparent);
             cells[y][x].shape.setOutlineThickness(1);
